# rlp: optimize big.Int decoding for size <= 32 bytes